refactor: CQRS 패턴으로 GroupRepository CRUD 분리
