+++ backend/src/main/java/com/woowacourse/momo/group/domain/GroupRepository.java
@@ -1,6 +1,10 @@
 package com.woowacourse.momo.group.domain;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.Repository;
 
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends Repository<Group, Long> {
+
+    Group save(Group group);
+
+    void deleteById(Long id);
 }
